refactor(avm): C++ sim cleanup and fixes (follow up) (#17361)
Follow up cleanup and fixes

+++ barretenberg/cpp/src/barretenberg/vm2/simulation/testing/mock_memory.hpp
@@ -15,7 +15,7 @@ class MockMemory : public MemoryInterface {
 
     MOCK_METHOD(void, set, (MemoryAddress index, MemoryValue value), (override));
     MOCK_METHOD(const MemoryValue&, get, (MemoryAddress index), (const, override));
-    MOCK_METHOD(uint32_t, get_space_id, (), (const, override));
+    MOCK_METHOD(uint16_t, get_space_id, (), (const, override));
 };
 
 } // namespace bb::avm2::simulation
